Add link to non-iData third party page from the cloud legal notices
As part of the autonomous operations project, we are now transitioning third party notices for cloud/CD away from iData process  (which publishes to the SAG website) and onto the C8Y non-iData process (which publishes pages under https://third-parties.c8y.io/ ). 

So far just one component is available on the new C8Y page (usagestats), but during the next few months more will be added one by one, That means from a customer perspective some components will be available in one place and others in the other, so we need links to both the C8Y and SAG pages until the full transition is complete.